+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 - The data displayed by the command-line client is now formatted as tables using the `terminal-table` gem
 - Added `--print-requests` option to the command-line client to see all the interactions with the IG Markets API
-- Added `--aggregate` option to `ig_markets positions list`
+- Added `--aggregate` option to `ig_markets positions list` to aggregate together positions with the same EPIC
+- Added `--instrument` option to `ig_markets transactions` to filter by instrument name
 - Support removal of limits and stops using `ig_markets positions update` and `ig_markets orders update`
 - Added `IGMarkets::Position#close_level`
 - Removed `IGMarkets::Position#formatted_size` and `IGMarkets#AccountTransaction#formatted_transaction_type`